+++ src/ska_low_mccs_spshw/tile/tile_poll_management.py
@@ -230,6 +230,7 @@ class RequestIterator:
         "IO",
         "DSP",
         "40G_PACKET_COUNT",
+        "POINTING_DELAYS",
     ]
 
     def __init__(self: RequestIterator):