+++ tests/unit/tile/test_tile_device.py
@@ -1004,6 +1004,7 @@ class TestMccsTile:
             "ddr_rd_cnt",
             "ddr_wr_cnt",
             "ddr_rd_dat_cnt",
+            "pointingDelays",
         ]
 
         all_excluded_attribute = (